Revert "LPS-99657 Class that implements ConfigurationBeanDeclaration should be in package *.settings.definition"
This reverts commit 5a7620b68147444ae7bf37abe6bb11ad0c1b2659.
