+++ modules/apps/site-navigation/site-navigation-breadcrumb-web/src/main/java/com/liferay/site/navigation/breadcrumb/web/internal/configuration/definition/SiteNavigationBreadcrumbPortletInstanceConfigurationBeanDeclaration.java
@@ -12,10 +12,10 @@
  * details.
  */
 
-package com.liferay.site.navigation.site.map.web.internal.settings.definition;
+package com.liferay.site.navigation.breadcrumb.web.internal.configuration.definition;
 
 import com.liferay.portal.kernel.settings.definition.ConfigurationBeanDeclaration;
-import com.liferay.site.navigation.site.map.web.internal.configuration.SiteNavigationSiteMapPortletInstanceConfiguration;
+import com.liferay.site.navigation.breadcrumb.web.internal.configuration.SiteNavigationBreadcrumbPortletInstanceConfiguration;
 
 import org.osgi.service.component.annotations.Component;
 
@@ -23,12 +23,12 @@ import org.osgi.service.component.annotations.Component;
  * @author Juergen Kappler
  */
 @Component(service = ConfigurationBeanDeclaration.class)
-public class SiteNavigationSiteMapPortletInstanceConfigurationBeanDeclaration
+public class SiteNavigationBreadcrumbPortletInstanceConfigurationBeanDeclaration
 	implements ConfigurationBeanDeclaration {
 
 	@Override
 	public Class<?> getConfigurationBeanClass() {
-		return SiteNavigationSiteMapPortletInstanceConfiguration.class;
+		return SiteNavigationBreadcrumbPortletInstanceConfiguration.class;
 	}
 
 }
