+++ modules/apps/site-navigation/site-navigation-site-map-web/src/main/java/com/liferay/site/navigation/site/map/web/internal/configuration/definition/SiteNavigationSiteMapPortletInstanceConfigurationBeanDeclaration.java
@@ -12,10 +12,10 @@
  * details.
  */
 
-package com.liferay.site.navigation.breadcrumb.web.internal.settings.definition;
+package com.liferay.site.navigation.site.map.web.internal.configuration.definition;
 
 import com.liferay.portal.kernel.settings.definition.ConfigurationBeanDeclaration;
-import com.liferay.site.navigation.breadcrumb.web.internal.configuration.SiteNavigationBreadcrumbPortletInstanceConfiguration;
+import com.liferay.site.navigation.site.map.web.internal.configuration.SiteNavigationSiteMapPortletInstanceConfiguration;
 
 import org.osgi.service.component.annotations.Component;
 
@@ -23,12 +23,12 @@ import org.osgi.service.component.annotations.Component;
  * @author Juergen Kappler
  */
 @Component(service = ConfigurationBeanDeclaration.class)
-public class SiteNavigationBreadcrumbPortletInstanceConfigurationBeanDeclaration
+public class SiteNavigationSiteMapPortletInstanceConfigurationBeanDeclaration
 	implements ConfigurationBeanDeclaration {
 
 	@Override
 	public Class<?> getConfigurationBeanClass() {
-		return SiteNavigationBreadcrumbPortletInstanceConfiguration.class;
+		return SiteNavigationSiteMapPortletInstanceConfiguration.class;
 	}
 
 }
